scheduler.rake outputs to logfile

+++ app/models/user.rb
@@ -131,18 +131,24 @@ class User < ActiveRecord::Base
 
   def self.remove_inactive_users
     inactive_users = User.where('confirmed <> ? and created_at < ?', true, 30.days.ago)
-    puts "Found #{inactive_users.count} inactive users\n"
-    puts "These users are about to be deleted:\n"
-    inactive_users.each_with_index do |u,i|
-      puts "#{i+1}. UserID: #{u.id}, Name: #{u.full_name}\n"
-    end
-    puts "Deleting...\n"
-    inactive_users.delete_all
-    if inactive_users.empty?
-      puts "Inactive users all deleted\n"
+    logger.info "Found #{inactive_users.count} inactive users\n"
+    if !inactive_users.empty?
+      logger.info "These users are about to be deleted:\n"
+      inactive_users.each_with_index do |u,i|
+        logger.info "#{i+1}. UserID: #{u.id}, Name: #{u.full_name}\n"
+      end
+      logger.info "Deleting...\n"
+      inactive_users.delete_all
+      if inactive_users.empty?
+        logger.info "Inactive users all deleted\n"
+      else
+        logger.info "Something must've gone wrong\n"
+      end
+    else
+      logger.info "No users to delete"
     end
 
-    puts "Finished task remove inactive users"
+    logger.info "Finished task remove inactive users"
 
   end
 end
